requirements.txt
dataclasses==0.6
PyYAML~=6.0
numpy==1.21.0
pandas==1.4.1
scikit-learn~=0.24.1
scipy==1.7.0
tqdm==4.56.2
jupyter

torch~=1.8.1+cu111
transformers>=4.11.0
torchmetrics~=0.6.2
pytorch-lightning~=1.5.10
pytorch_nlp==0.5.0

ray~=2.0.0.dev0
mlflow==1.14.1

# if pip install doesn't work for ray tune, separately run the below command
# pip install "ray[default] @ https://s3-us-west-2.amazonaws.com/ray-wheels/latest/ray-2.0.0.dev0-cp38-cp38-manylinux2014_x86_64.whl"
